Project: Rename files, refactor code and create aurora data file
