+++ src/components/Project/StyledProjects.js
@@ -1,7 +1,4 @@
-import React, { Component } from 'react'
-import PropTypes from 'prop-types'
 import styled from 'styled-components'
-// import { Link } from 'react-static'
 
 const ProjectTemplate = styled.div`padding-top: 100px; text-align: left;`
 
@@ -15,18 +12,8 @@ const Title = styled.h1`
   color: black;
   @media (min-width: 768px) {
     font-size: 55px;
-    &:before {
-      content: '';
-      position: absolute;
-      top: 50%;
-      background-color: #000;
-      transform: translateY(-50%);
-      right: -120px;
-      height: 10px;
-      width: 100px;
-    }
   }
-`;
+`
 
 const Banner = styled.div`
   height: 40vh;
@@ -40,7 +27,7 @@ const Banner = styled.div`
   @media (min-width: 768px) {
     height: 60vh;
   }
-`;
+`
 
 const InfoContainer = styled.div`
   width: 100%;
@@ -67,7 +54,7 @@ const InfoContainer = styled.div`
       flex: 0 1 50%;
     }
   }
-`;
+`
 
 const DetailList = styled.ul`
   padding: 0;
@@ -111,7 +98,7 @@ const DetailList = styled.ul`
       margin-top: 30px;
     }
   }
-`;
+`
 
 const ButtonGradient = styled.a`
   background: linear-gradient(-134deg, #3fdfa4 0%, #3e45b9 100%);
@@ -124,7 +111,7 @@ const ButtonGradient = styled.a`
   @media (max-width: 768px) {
     display: none;
   }
-`;
+`
 
 const VideoContent = styled.div`
   display: flex;
@@ -150,7 +137,7 @@ const VideoContent = styled.div`
     height: auto;
     z-index: 2;
   }
-`;
+`
 
 const MobileContainer = styled.div`
   padding: 100px 0;
@@ -184,86 +171,13 @@ const MobileContainer = styled.div`
   }
 `
 
-class Template extends Component {
-  componentDidMount () {
-    window.scrollTo(0, 0)
-  }
-  render () {
-    const { 
-      paragraph,
-      title,
-      details,
-      videoSrc,
-      buttonLink,
-      iphone,
-    } = this.props
-    const bannerImage = { backgroundImage: `url('${this.props.banner}')` }
-
-    return (
-      <ProjectTemplate>
-        <div className="container">
-          <Title>{title}</Title>
-        </div>
-        <Banner style={bannerImage} />
-        <InfoContainer>
-          <div className="flex-item flex-half">
-            <p dangerouslySetInnerHTML={{ __html: paragraph }} />
-          </div>
-          <div className="flex-item flex-half">
-            <DetailList>
-              {
-                details.map((detail, index) => (
-                  <li key={index}>
-                    <p className="title">{detail.title}</p>
-                    <p
-                      className="content"
-                      dangerouslySetInnerHTML={{ __html: detail.content }}
-                    />
-                  </li>
-                ))
-              }
-            </DetailList>
-          </div>
-        </InfoContainer>
-        <ButtonGradient href={buttonLink} target="_blank">
-          View Website
-        </ButtonGradient>
-        <VideoContent>
-          <div className="img-wrapper">
-            <video loop muted playsInline autoPlay>
-              <source src={videoSrc} type="video/mp4" />
-              Your browser does not support the video tag.
-            </video>
-            <img src="/assets/img/desktop.svg" alt="Desktop" />
-          </div>
-        </VideoContent>
-        <MobileContainer>
-          {
-            iphone.map((phone, index) => (
-              <div className="mobile-wrapper" key={index}>
-                <img className="screenshot" src={phone} alt="" />
-                <img
-                  className="iphone"
-                  src="/assets/img/iphone.svg"
-                  alt="iPhone"
-                />
-              </div>
-            ))
-          }
-        </MobileContainer>
-      </ProjectTemplate>
-    )
-  }
+export {
+  ProjectTemplate,
+  Title,
+  Banner,
+  InfoContainer,
+  DetailList,
+  ButtonGradient,
+  VideoContent,
+  MobileContainer,
 }
-
-Template.propTypes = {
-  title: PropTypes.string,
-  banner: PropTypes.string,
-  paragraph: PropTypes.string,
-  details: PropTypes.array,
-  buttonLink: PropTypes.string,
-  videoSrc: PropTypes.string,
-  iphone: PropTypes.array,
-}
-
-export default Template
