+++ src/data/auroraProjects.js
@@ -1,0 +1,15 @@
+const auroraMedical = {
+
+}
+
+const dailySpecial = {
+
+}
+
+const auroraEurope = {
+
+}
+
+export {
+  auroraMedical
+}
