+++ src/data/homeProjects.js
@@ -21,10 +21,19 @@ export default [
     src: '/assets/img/home/auroraeurope-project.jpg',
     name: 'Aurora Europe',
     link: '/aurora-europe',
-    year: '2019',
+    year: '2020',
     company: 'Aurora',
     tech: 'React',
     role: 'Manager',
+  },
+  {
+    src: '/assets/img/home/acs-project.jpg',
+    name: 'Aurora Cannabis Store',
+    link: '/aurora-cannabis-store',
+    year: '2019',
+    company: 'Aurora',
+    tech: 'React',
+    role: 'Developer',
   },
   {
     src: '/assets/img/home/investor-project.jpg',
