+++ src/pages/Gryphon.js
@@ -1,12 +1,12 @@
 import React, { Component } from 'react'
-import Template from 'components/Projects/Template'
+import Project from '../components/Project'
 import { gryphon } from '../data/bamProjects'
 
 export default class GryphonPage extends Component {
   render () {
     return (
       <section>
-        <Template
+        <Project
           title={gryphon.title}
           banner={gryphon.banner}
           paragraph={gryphon.paragraph}
