+++ src/pages/Home.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react'
 import styled from 'styled-components'
-import Hero from 'components/Home/Hero'
-import Quote from 'components/Home/Quote'
-import Projects from 'components/Home/Projects'
+import HeroSection from 'components/Home/HeroSection'
+import QuoteSection from 'components/Home/QuoteSection'
+import ProjectsSection from 'components/Home/ProjectsSection'
 
 const HomeWrapper = styled.section``
 
@@ -13,9 +13,9 @@ export default class Home extends Component {
   render () {
     return (
       <HomeWrapper class="loader">
-        <Hero />
-        <Quote />
-        <Projects />
+        <HeroSection />
+        <QuoteSection />
+        <ProjectsSection />
       </HomeWrapper>
     )
   }
